Added webpack to the mix. Reorganized files & folders.
