+++ .gitignore
@@ -1,4 +1,2 @@
 .DS_Store
 node_modules
-app.js
-example.html
