+++ example/example.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>Fullscreen Background Image Gallery</title>
+  <link rel="stylesheet" href="../dist/styles.css">
+</head>
+
+<body>
+  <script src="../dist/big.js"></script>
+  <script src="./app.js"></script>
+</body>
+</html>
