+++ package.json
@@ -4,7 +4,7 @@
   "description": "A JavaScript background gallery that fades from one image to the next.",
   "main": "BGimageGallery-npm.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -17,9 +17,17 @@
     "carousel"
   ],
   "author": "Qodesmith",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/qodesmith/background-image-gallery/issues"
   },
-  "homepage": "https://github.com/qodesmith/background-image-gallery#readme"
+  "homepage": "https://github.com/qodesmith/background-image-gallery#readme",
+  "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "babel-loader": "^8.0.2",
+    "randomize-array": "^1.0.0",
+    "webpack": "^4.18.0",
+    "webpack-cli": "^3.1.0"
+  }
 }
